+++ R/svyplan-package.R
@@ -1,7 +1,8 @@
 #' @keywords internal
 "_PACKAGE"
 
-#' @importFrom graphics hist
+#' @importFrom graphics abline barplot hist legend points
 #' @importFrom grDevices nclass.FD
 #' @importFrom stats lm.fit optim optimize pnorm quantile qnorm uniroot var weighted.mean
+#' @importFrom utils modifyList
 NULL